Creature

Description

Documentation

getCreatureInfo

Arguments

:: Creature
Input: Creature to get info from

-> String
Output: All info about the given Creature in one String

getHp

Arguments

:: Creature
Input: Creature to get hp level from

-> Int
Output: Hp level of given Creature

Retrieve hp level of given Creature

alterHp

Arguments

:: Int
Input: Number of hp to add to current hp level, can be positive or negative

-> Creature
Input: Creature to alter hp for

-> Creature
Output: Given Creature with altered hp level

Increase or decrease hp of given Creature

moveCreature

Arguments

:: Creature
Input: Creature to change position for

-> Direction
Input: Direction to move to. Can be Up', Down', Left', Right'

-> Creature
Output: Given Creature with +1 in given Direction

Replace the position of a creature based on the given Direction

data Creature

A Creature will walk through the grid searching for food, trying not to die.

getCreaturePos

Arguments

:: Creature
Input: Creature to get position from

-> Pos
Output: Position in (x,y) format

Returns positions of given Creature

generateCreatures

Arguments

:: Int
Input: Number of Creatures to create

-> Int
Input: Highest allowed position in the grid

-> StdGen
Input: Seed 1 (For x position)

-> StdGen
Input: Seed 2 (For y position)

-> [Creature]
Output: Newly generated Creatures

generate creatures with a pseudo-random position using two seeds