Helpers

Description

Documentation

readUntilNumeric

Arguments

:: IO String
String to return if all characters are numeric

Get text from user until input is all numeric

replace

Arguments

:: Eq a

=> a
Input: Value to look for

-> a
Input: Value to replace it with

-> [a]
Input: Input list

-> [a]
Output: Output list

Basic replace function

exitWithMessage

Arguments

:: String
Input: Message to show to user

-> IO a
Output: Message printed to console + Exited program

Exit program with message

randomList

Arguments

:: Int
Input: Number of Integers to be generated

-> Int
Input: Highest allowed value in list

-> StdGen
Input: Seed

-> [Int]
Output: List of pseudo-random generated Integers

Create pseudo-random list of n numbers using a seed

getRandomSeed

Arguments

:: Int
Input: Number to use as seed for random seeds

-> Int
Input: index of element to take from list

-> StdGen
Output: seed

Function to generate multiple random seed from one number