Position

Description

Documentation

getDistance

Arguments

:: Pos
Input: Position 1

-> Pos
Input: Position 2

-> Int
Output: Absolute distance between two Positions

Get the distance between two Positions

changePos

Arguments

:: Pos
Input: Position to change

-> Int
Input: Axis/index to change

-> Int
Input: Number to add (positive or negative) to regarding index

-> Pos
Output: Given position changed based on axis and number

Alter the position on the given axis with a given value

generateRandomPositions

Arguments

:: Int
Input: Number of pseudo-random Positions to be generated

-> Int
Input: Highest allowed value in to be generated

-> StdGen
Input: Seed 1 (For x position)

-> StdGen
Input: Seed 2 (For y position)

-> [Pos]
Output: List of pseudo-random generated Positions

Create list of pseudo-random Positions using two seeds

decideDirection

Arguments

:: Pos
Input: Current position

-> Pos
Input: Target position

-> Direction
Output: Direction based on decision making

Pathfinding to decide a direction (Up',Down',Right',Left') based on current and target position

type Pos = (Int, Int)

Alternative notation of (Int,Int) to make code more readable

data Direction

datatype to make Directions more readable and understandable according to the grid.

Constructors

Up'

Down'

Left'

Right'